feat: use LuaInstancesCAS as LuaInstances impl
