+++ example/src/jmh/java/party/iroiro/luajava/jmh/instances/LuaInstancesReaderBenchmark.java
@@ -4,7 +4,6 @@ import org.openjdk.jmh.annotations.*;
 import party.iroiro.luajava.luajit.LuaJit;
 
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 @Fork(1)
 @Warmup(iterations = 3)
@@ -14,22 +13,11 @@ import java.util.concurrent.atomic.AtomicBoolean;
 @State(Scope.Thread)
 public class LuaInstancesReaderBenchmark {
 
-    private static final AtomicBoolean isSetup = new AtomicBoolean(false);
-    @Param({"synchronized", "AtomicReference", "CopyOnWrite", "CAS"})
-    @SuppressWarnings("NotNullFieldNotInitialized")
-    public String instancesType;
-
     @SuppressWarnings("NotNullFieldNotInitialized")
     private LuaJit L;
 
     @Setup
     public void setup() {
-        synchronized (isSetup) {
-            if (!isSetup.getAndSet(true)) {
-                LuaInstancesAccess.setInstances(instancesType);
-            }
-        }
-
         L = new LuaJit();
         L.run("function met_call() for _ = 1, 1000 do java.luaify() end end");
     }
